+++ minisound/example/lib/main.dart
@@ -22,10 +22,7 @@ class _ExamplePageState extends State<ExamplePage> {
   final engine = Engine();
   final recorder = Recorder();
 
-  late final initFuture = Future.wait([
-    engine.init(),
-    // recorder.init(),
-  ]);
+  late final initFuture = engine.init().then((_) => recorder.init());
 
   @override
   Widget build(BuildContext context) {