+++ minisound_ffi/lib/minisound_ffi_bindings.dart
@@ -167,11 +167,11 @@ class MinisoundFfiBindings {
 
   void noise_sound_data_set_seed(
     ffi.Pointer<NoiseSoundData> self,
-    int seed,
+    int value,
   ) {
     return _noise_sound_data_set_seed(
       self,
-      seed,
+      value,
     );
   }
 
@@ -712,6 +712,120 @@ class MinisoundFfiBindings {
   late final _engine_generate_pulse = _engine_generate_pulsePtr.asFunction<
       int Function(ffi.Pointer<Engine>, ffi.Pointer<Sound>, double, double)>();
 
+  int sizeof_recording() {
+    return _sizeof_recording();
+  }
+
+  late final _sizeof_recordingPtr =
+      _lookup<ffi.NativeFunction<ffi.Size Function()>>('sizeof_recording');
+  late final _sizeof_recording =
+      _sizeof_recordingPtr.asFunction<int Function()>();
+
+  ffi.Pointer<ffi.Uint8> recording_get_buf(
+    ffi.Pointer<Recording> self,
+  ) {
+    return _recording_get_buf(
+      self,
+    );
+  }
+
+  late final _recording_get_bufPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Pointer<ffi.Uint8> Function(
+              ffi.Pointer<Recording>)>>('recording_get_buf');
+  late final _recording_get_buf = _recording_get_bufPtr
+      .asFunction<ffi.Pointer<ffi.Uint8> Function(ffi.Pointer<Recording>)>();
+
+  int recording_get_size(
+    ffi.Pointer<Recording> self,
+  ) {
+    return _recording_get_size(
+      self,
+    );
+  }
+
+  late final _recording_get_sizePtr =
+      _lookup<ffi.NativeFunction<ffi.Size Function(ffi.Pointer<Recording>)>>(
+          'recording_get_size');
+  late final _recording_get_size =
+      _recording_get_sizePtr.asFunction<int Function(ffi.Pointer<Recording>)>();
+
+  ffi.Pointer<RecorderBuffer> recorder_buffer_alloc() {
+    return _recorder_buffer_alloc();
+  }
+
+  late final _recorder_buffer_allocPtr =
+      _lookup<ffi.NativeFunction<ffi.Pointer<RecorderBuffer> Function()>>(
+          'recorder_buffer_alloc');
+  late final _recorder_buffer_alloc = _recorder_buffer_allocPtr
+      .asFunction<ffi.Pointer<RecorderBuffer> Function()>();
+
+  int recorder_buffer_init(
+    ffi.Pointer<RecorderBuffer> self,
+    int encoding,
+    ffi.Pointer<ffi.Void> v_device,
+  ) {
+    return _recorder_buffer_init(
+      self,
+      encoding,
+      v_device,
+    );
+  }
+
+  late final _recorder_buffer_initPtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Int32 Function(ffi.Pointer<RecorderBuffer>, ffi.Int32,
+              ffi.Pointer<ffi.Void>)>>('recorder_buffer_init');
+  late final _recorder_buffer_init = _recorder_buffer_initPtr.asFunction<
+      int Function(ffi.Pointer<RecorderBuffer>, int, ffi.Pointer<ffi.Void>)>();
+
+  void recorder_buffer_uninit(
+    ffi.Pointer<RecorderBuffer> self,
+  ) {
+    return _recorder_buffer_uninit(
+      self,
+    );
+  }
+
+  late final _recorder_buffer_uninitPtr = _lookup<
+          ffi.NativeFunction<ffi.Void Function(ffi.Pointer<RecorderBuffer>)>>(
+      'recorder_buffer_uninit');
+  late final _recorder_buffer_uninit = _recorder_buffer_uninitPtr
+      .asFunction<void Function(ffi.Pointer<RecorderBuffer>)>();
+
+  int recorder_buffer_write(
+    ffi.Pointer<RecorderBuffer> self,
+    ffi.Pointer<ffi.Uint8> data,
+    int data_len_pcm,
+  ) {
+    return _recorder_buffer_write(
+      self,
+      data,
+      data_len_pcm,
+    );
+  }
+
+  late final _recorder_buffer_writePtr = _lookup<
+      ffi.NativeFunction<
+          ffi.Int32 Function(ffi.Pointer<RecorderBuffer>,
+              ffi.Pointer<ffi.Uint8>, ffi.Size)>>('recorder_buffer_write');
+  late final _recorder_buffer_write = _recorder_buffer_writePtr.asFunction<
+      int Function(ffi.Pointer<RecorderBuffer>, ffi.Pointer<ffi.Uint8>, int)>();
+
+  Recording recorder_buffer_consume(
+    ffi.Pointer<RecorderBuffer> self,
+  ) {
+    return _recorder_buffer_consume(
+      self,
+    );
+  }
+
+  late final _recorder_buffer_consumePtr = _lookup<
+          ffi.NativeFunction<Recording Function(ffi.Pointer<RecorderBuffer>)>>(
+      'recorder_buffer_consume');
+  late final _recorder_buffer_consume = _recorder_buffer_consumePtr
+      .asFunction<Recording Function(ffi.Pointer<RecorderBuffer>)>();
+
   ffi.Pointer<Recorder> recorder_alloc() {
     return _recorder_alloc();
   }
@@ -851,6 +965,19 @@ final class Sound extends ffi.Opaque {}
 
 final class Engine extends ffi.Opaque {}
 
+final class Recording extends ffi.Struct {
+  external ffi.Pointer<ffi.Uint8> buf;
+
+  @ffi.Size()
+  external int size;
+}
+
+final class RecorderBuffer extends ffi.Opaque {}
+
+abstract class RecordingEncoding {
+  static const int RECORDING_ENCODING_WAV = 1;
+}
+
 final class Recorder extends ffi.Opaque {}
 
 abstract class RecorderFormat {
@@ -859,15 +986,4 @@ abstract class RecorderFormat {
   static const int RECORDER_FORMAT_S24 = 3;
   static const int RECORDER_FORMAT_S32 = 4;
   static const int RECORDER_FORMAT_F32 = 5;
-}
-
-abstract class RecordingEncoding {
-  static const int RECORDING_ENCODING_WAV = 1;
-}
-
-final class Recording extends ffi.Struct {
-  external ffi.Pointer<ffi.Uint8> buf;
-
-  @ffi.Size()
-  external int size;
 }