+++ minisound_ffi/src/include/minisound.h
@@ -2,7 +2,9 @@
 #define MINISOUND_H
 
 #include "engine.h"
-#include "recorder.h"
+#include "recorder/recorder.h"
+#include "recorder/recorder_buffer.h"
+#include "recorder/recording.h"
 #include "sound.h"
 #include "sound_data/encoded_sound_data.h"
 #include "sound_data/noise_sound_data.h"